wasm3.0: Advance Wasm 3.0 GC support and add a fast release gate
- route struct.new through payload-backed preserved helpers so arbitrary field
  counts work across the native backends
- finish GC array decode/lowering/runtime plumbing and tighten related machine
  IR, validator, and preserved-helper handling
- fix arm64 preserved-helper trap propagation so post-call helper errors are not
  lost on the error path
- add targeted native regressions for array OOB and wide struct construction
- update spectest discovery/expectations for the GC work that is now live

+++ sf-nano-core/src/vm/wasm/sir/primitive_op.rs
@@ -201,7 +201,7 @@ macro_rules! for_each_primitive_op {
             ExternConvertAny => (1, 1),
             RefTest { ref_type: crate::value_type::RefType } => (1, 1),
             RefCast { ref_type: crate::value_type::RefType } => (1, 1),
-            StructNew { type_idx: u32, field_count: u8 } => (usize::from(*field_count), 1),
+            StructNew { type_idx: u32, field_count: u32 } => (*field_count as usize, 1),
             StructNewDefault { type_idx: u32 } => (0, 1),
             StructGet { type_idx: u32, field_idx: u32 } => (1, 1),
             StructGetS { type_idx: u32, field_idx: u32 } => (1, 1),
